requirements.txt
numpy
torch>=1.7.0
torchvision>=0.8.0
Pillow
scikit-image 
kornia
tqdm
pandas
compressai
bm3d
diffusion
